Start implementing the SpotlightWizard
This patch enables the Spotlight wizard and adds the transitions,
backdrop, and modal fields to it.

Some reorganization was done with respect to the `*FormData` and
`*MessageContent` types in order to make them more correllated with
their components.

The serializer has been updated to serialize the fields we have
provided.