Add Sentry integration
The Sentry integration can be tested locally by setting the `SENTRY_DSN`
and `SENTRY_RELEASE` environment variables.

Content blockers such as uBlock Origin and Ghostery block connections to
Sentry's ingest server, so we try to detect these errors and show a
toast notification asking the user to disable them when we cannot
connect to the ingest server.