Correctly focus hidden invalid controls
Previously, if a control failed validation but was not visible due to
being a child of a `<TabbedInput>` control whose tab is not active,
focusing would fail. Form validation usually focuses the first control
that is invalid, but when the control isn't visible they cannot be
focused.

The way react-hook-form focuses inputs is that `register()` returns a
ref callback. To work around this behaviour, we provide our own
`register()` function that wraps the library-provided `register()`
function but returns a ref callback that calls the original ref callback
with the ref wrapped in a `Proxy`. When `focus()` is called on the
proxy, the tab will first be activated before the actual focus call.

This results in focusing working correctly when invalid controls are not
visible.