Make LocalizableTextInput generic over form data
`<LocalizableTextInput>` no longer requires the form data to be
`WizardFormData`, which will make writing unit tests much simpler.

Due to type inference limits in TypeScript, we cannot concatenate
strings to safely build `Path<T>` pointing *inside* a
`LocalizableTextFormData` inside a generic form data type. However, we
know that concatenating a `Path<T>` to a `LocalizableTextFormData` and a
field name of the same will result in a valid `Path<T>`. The `fieldname`
helper function added does these assertions and returns a `Path<T>`,
which is much safer than doing concatenation in place because it
enforces that the field name is valid.