Deserialize rich localizable text with presets
If the field is set to support rich text, then the presence of any rich
text property will cause the text to be deserialized with that preset.
Currently the input only supports plain text and a single rich text
preset, so it will need to gain support for individual rich text
properties to fully deserialize rich text correctly.

The remeda library was added as a functional utility library.