Keep track of ref proxies in TabbedInput
Previously we were re-creating a new `Proxy` on each call to `render()`,
resulting in React and react-hook-form thinking that the underlying
element had changed. This lead to an accumulation of refs in
react-hook-form's internal state for checkbox fields inside
`<TabbedInput>`s. This broke `<LocalizableTextInput>` on the
`<InfoBarButtonsInput>` and on the `<SpotlightScreensInput>`.

Now we cache our proxies in a `WeakMap` indexed by the DOM element so
that we will compare equal between renders as long as the element hasn't
changed.

A unit test has been added to ensure that `<LocalizedTextInput>` works
correctly inside a `<TabbedInput>`. The test fails without the fix in
this patch.

Fixes #56.