Allow empty triggers when not required
The trigger field is now validated with a function that allows for
empty string inputs in the case where we do not want to set a trigger on
the message and the messaging surface in question does not require it.

Likewise, while serializing the trigger field, we will not attempt to
parse it as JSON if it is empty.